+++ src/extension.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { createMacro } from './commands/createMacro';
 import { debugMacro } from './commands/debugMacro';
-import { getActiveMacroEditor } from './commands/getActiveMacroEditor';
+import { getActiveMacroEditorUri } from './commands/getActiveMacroEditorUri';
 import { runMacro } from './commands/runMacro';
 import { selectMacroFile } from './commands/selectMacroFile';
 import { showRunningMacros } from './commands/showRunningMacros';
@@ -24,27 +24,27 @@ export async function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     manager,
     new StatusBarItem(manager),
-    vscode.commands.registerCommand('macros.debug', async () => {
-      const uri = await selectMacroFile();
+    vscode.commands.registerCommand('macros.debug', async (pathOrUri?: string | vscode.Uri) => {
+      const uri = await getMacroUri(pathOrUri);
       if (uri) {
         await debugMacro(manager, uri);
       }
     }),
     vscode.commands.registerCommand('macros.debug.activeEditor', async () => {
-      const uri = await getActiveMacroEditor();
+      const uri = await getActiveMacroEditorUri();
       if (uri) {
         await debugMacro(manager, uri);
       }
     }),
     vscode.commands.registerCommand('macros.new.macro', () => createMacro(context)),
-    vscode.commands.registerCommand('macros.run', async () => {
-      const uri = await selectMacroFile();
+    vscode.commands.registerCommand('macros.run', async (pathOrUri?: string | vscode.Uri) => {
+      const uri = await getMacroUri(pathOrUri);
       if (uri) {
         await runMacro(manager, uri);
       }
     }),
     vscode.commands.registerCommand('macros.run.activeEditor', async () => {
-      const uri = await getActiveMacroEditor();
+      const uri = await getActiveMacroEditorUri();
       if (uri) {
         await runMacro(manager, uri);
       }
@@ -52,5 +52,11 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand('macros.run.mru', () => runMacro(manager, mruMacro)),
     vscode.commands.registerCommand('macros.run.show', () => showRunningMacros(manager)),
   );
+
+  async function getMacroUri(pathOrUri?: string | vscode.Uri) {
+    return pathOrUri
+      ? (pathOrUri instanceof vscode.Uri ? pathOrUri : vscode.Uri.parse(pathOrUri))
+      : await selectMacroFile();
+  }
 }