+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.7
+- `[Run|Debug] Macro…` commands take in an optional `pathOrUri` argument.
+- Add `Rerun` option on error.
+
 ## 0.0.6
 - `[Run|Debug] Macro…` commands subsume `[Run|Debug] Macro (Source Directories)` commands to simplify UX.